fix: remove unused context passed in at tool creation
This change cleans up an extra context passed in at tool creation, which shadows the handler context. In practice each handler runs with a context which is scoped to the stdio server.

To remove all this confusion I am removing this context, and to resolve log propegation via the handlers context I have introduced an option to set it up correctly, which fixes the logger. :facepalm: